move PageNotFoundException to internal.exception package
